完了ステータス入力コンポーネントを作成した
